requirements.txt
fastapi
uvicorn
langchain[openai]
openai
python-dotenv
redis
psycopg2